requirements.txt
asgiref==3.11.0
black==25.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
Django==5.2.8
djangorestframework==3.16.1
idna==3.11
isort==7.0.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.5.0
pytokens==0.3.0
requests==2.32.5
ruff==0.14.5
sqlparse==0.5.3
urllib3==2.5.0
